+++ read_lat.c
@@ -380,7 +380,7 @@ static void print_report(struct perftest_parameters *user_param) {
 	}
 
 	median = get_median(user_param->iters - 1, delta);
-	printf(REPORT_FMT_LAT,user_param->size,user_param->iters,delta[0] / cycles_to_units ,
+	printf(REPORT_FMT_LAT,(unsigned long)user_param->size,user_param->iters,delta[0] / cycles_to_units ,
 	       delta[user_param->iters - 2] / cycles_to_units ,median / cycles_to_units );
 
 	free(delta);
